requirements.txt
openai==1.61.1
python-telegram-bot==21.10
Requests==2.32.3
smolagents==1.7.0
eventregistry==9.1